files task1.rb and task2.rb - deleted. They were replaced by three files: games_methods, mario_game and tanks_game. Functionality need fixes
